Adding and Adjusting Article views and controllers
