+++ app/controllers/authors_controller.rb
@@ -1,0 +1,6 @@
+class AuthorsController < ApplicationController
+  before_action :authenticate_author!
+
+  layout 'authors'
+  
+end
